We now have an application-supervised supervisor for the Game.
Figured out some of my cygwin headaches on the way, related to a dreaded
UnicodeConversionError due to a chocolatey update; my fix is to backup+rename
mix.bat and mix.exe in the chocolatey/bin/ directory and always call "mix.bat"
instead of "mix".  Now I can easily run:

  iex --werl -S mix  # or just iex -S mix on my other systems

and stay in there to dynamically load changed modules and debug.  Is there a
way to convince iex to automatically load changed files?  I know this is
something done in Phoenix using brunch and elm-brunch.

I have been using the observer, which has been super helpful:

  :observer.start

With these cascading supervisors in place, I now have this setup, as far as I
understand it:

  App -> NumberGuess -> Sup -> GameSup -> Game

  App         = application_master (called via proc_lib:init_p/5)
  NumberGuess = application_master (called via application_master:start_it/4)
  Sup         = Elixir.NumberGuess.Supervisor
  GameSup     = Elixir.NumberGuess.GameSup
  Game        = Elixir.NumberGuess.Game (my worker)

This looks similar to Elixir.Logger's setup, so I think I am on the right
track.