Alternative approach to working with tested child servers
I removed the custom stop, which sat a bit wrong with me.  Updated the test to
use Process.exit/2.  I am not sure what is better in a test of a server: link
the health of the child to the test? or suffer the timeout wait for a call to
finish with bad results?