The main Game can now die without losing state.
This implements the technique Dave Thomas revealed in Programming Elixir by
capturing state in a side stash and arranging a supervisor tree as:

  Supervisor -> DB
             -> Supervisor -> Game

After thinking and discussing with others, this is a great pattern.  Game
depends on DB, so we should start DB first.  If the Game goes down, we restart
with DB's current knowledge.  If DB goes down, we restart it all because we
can no longer necessarily trust what is in the Game.

Although I accomplished the task, I am a little concerned about some things...

For example, there is the funky management of the state.  I felt like I was
reimplementing the same data structure construction in multiple places.  I
should probably include the db_pid as part of the fundamental state.

I am very much slipping on documentation in the excitement to get some working
code out.  Dialyzer has been helpful, but I am definitely missing a number of
specs for clarity.

The directory hierarchy and class names feel off with the new supervisor
hierarchy.  I think I should have taken the generic concept of Game and split
it into Engine vs DB with a Game supervisor instead of Game vs DB with a DB
supervisor.

Aside from future refactoring, we are close to having an interface for humans
to play this game instead of sending messages through iex.