Moved new state logic from the DB to the Game.
I realized it was silly that the generic DB had so much knowledge, when those
calculations should be done as part of the "Game".

While I was at it, I also switched the setting of state from a call to a cast;
we are terminating, and there is nothing we care about from the response.