Unified the game state.
Now the db_pid is stored as part of the state, which removes the awkward state
tuple with inner %State map pattern.  I have mixed feelings about
this, since the db_pid is stored in a structure that is maintained by the
process with that pid, but the original situation seemed more awkward and
involved more {}s.