Added specs and docs to external APIs.
As a result of clearly defining type expectations, I updated tests so that
they could help me validate a few returns more clearly.  Granted, it is a
little silly to verify that GenServer.cast returns :ok, but since I am
casting, I might as well throw that in there for code reading clarity.