Renamed and moved modules around to a better? structure.
Documented the supervisor hierarchy in the NumberGuess application.  The
application no longer builds an anonymous supervisor but instead follows the
example/advice at

  http://elixir-lang.org/docs/stable/elixir/Application.html

to call a separate Supervisor's start_link().  It just seems clearer than the
previous approach.  I do not understand the tradeoffs here; is it a difference
of opinion between the authors of the mix template and the elixir-lang doc
authors?  But the awkward one-function import is gone now.