Dialyzer was mad at me.  It helped me refactor a few things.
I am a little annoyed by

  Guard test is_binary(...) can never succeed

in the middle of string interpolation.  I had to conjure the awkward

  "Text #{some_number_var |> Interger.to_string} text"

solution to make dialyzer happy.  I would think that it is already obvious
that the intent is to use the string representation of the variable, and it
does so despite dialyzer's messaging.