Converted DB from a GenServer to an Agent.
Excellent code review suggestion from mgwidmann at

  https://www.reddit.com/r/elixir/comments/3vr27v/code_review_my_attempt_at_a_console_number_guess/

This helped nail down what Agents do for me.  Basically, this abstraction
removed two handlers and a behavior.

I really need to find a good example of a unit test that communicates with
GenServers.  I had no idea that the test was run with the entire application
environment in play; that seems convenient, but simultaneously, it means I am
not in control of the environment setup, and it's global.  That's bad for
testing.  I am probably missing something here.