Reduced handle_call param binding and reply clutter.
warnt on reddit made a couple suggestions with regard to how to handle map
bindings and return values.  The alternative map approach does make the code
more readable, but now I have run into an infamous dialyzer+elixir bug:

  engine.ex:71: The call _@1:'guesses'() requires that _@1 is of type atom() | tuple() not #{}
  engine.ex:72: The call _@2:'db_server'() requires that _@2 is of type atom() | tuple() not #{}
  engine.ex:77: The call _@1:'number'() requires that _@1 is of type atom() | tuple() not #{}
  engine.ex:78: The call _@2:'db_server'() requires that _@2 is of type atom() | tuple() not #{}

Bug identified here:

  https://github.com/elixir-lang/elixir/issues/2478

Jose's remark on this was "The majority of remaining warnings can only be
fixed if Elixir has a better idea of types in the system and this is something
we will tackle on after 1.0."  I am using 1.1.1 at the moment.  Based on this
comment, I thought that maybe adding a @type for State and a @spec for
handle_call/3 would solve the problem, but it did not.

The other concern I have with this change is that warnt's suggestions for
reply simplification violate an idiom I saw throughout my reading of Erlang
books: shouldn't returned values include some typing information?  I am
willing to run with the notion of not parroting a request back, but at least

  {:ok, number}

is more preferable to

  number

right?

Nonetheless, the exercise is good.